// react-ui/types/generated/models/hex-position.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export type HexPosition = 'Right' | 'BottomRight' | 'BottomLeft' | 'Left' | 'TopLeft' | 'TopRight' | 'None';

export const HexPosition = {
    Right: 'Right',
    BottomRight: 'BottomRight',
    BottomLeft: 'BottomLeft',
    Left: 'Left',
    TopLeft: 'TopLeft',
    TopRight: 'TopRight',
    None: 'None',
} as const;